fix(barefoot): change time field type from int to float in _StateMessage
Update the type hint for the time field in the _StateMessage TypedDict to float. This change ensures the timestamp can represent sub-second precision for the last sample, which is often required for high-frequency tracking data.

+++ mmlib/matcher/online/barefoot/_types.py
@@ -12,7 +12,7 @@ class _Candidate(TypedDict):
    
 class _StateMessage(TypedDict):
     id: str
-    time: int       # Timestamp da última amostra
+    time: float       # Timestamp da última amostra
     point: str      # WKT POINT (amostra original)
     osm_id: int
     path_osm_ids: list[int]
